Lab5/TODO/requirements.txt
Django~=3.1.6
djangorestframework-jwt~=1.11.0
djangorestframework~=3.12.2